Fix wake from sleep missed messages
On some systems the websocket does not get disconnected when the system goes to sleep. In this case, message received during sleep are not shown to the user.
This commit introduces a timer which set a timestamp on an interval. When the time  elapsed since the last timestamp reaches a limit, the listener will disconnect and reconnect, such that missed messages are retreived by the application. This timeout will only occur after the system wakes up from sleep.